requirements.txt
Django~=3.2.12
djangorestframework~=3.13.0
gunicorn
psycopg2-binary~=2.8.6
coverage
